requirements.txt
python-dotenv == 1.0.1
langchain-openai == 0.3.3
langchain-core == 0.3.33
python-telegram-bot == 21.10